requirements.txt
# Core dependencies
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
torch==2.0.1
torchvision==0.15.2

# Reinforcement Learning
stable-baselines3==2.3.0
gymnasium==0.29.1
shimmy==0.2.1

# Trading and Technical Analysis
binance-futures-connector==3.3.1
pandas-ta==0.3.14b
ta==0.10.2
ccxt==4.1.0
websocket-client==1.6.1

# Performance optimization
numba==0.57.1
pyarrow==12.0.1

# Machine Learning
lightgbm==4.0.0
scikit-learn==1.3.0
shap==0.42.1
matplotlib==3.7.2
seaborn==0.12.2

# Google Cloud Platform
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.36.0
google-cloud-logging==3.5.0
google-cloud-monitoring==2.15.1

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
aiohttp==3.8.5

# Monitoring and Observability
prometheus-client==0.18.0
grafana-api==1.0.3
structlog==23.1.0

# Database and Caching
redis==4.6.0
asyncpg==0.28.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.2
tqdm==4.66.1
pyyaml==6.0.1
click==8.1.7
tenacity==8.2.3

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
black==23.10.1
ruff==0.1.5
mypy==1.6.1
isort==5.12.0
pre-commit==3.5.0

# Security
bandit==1.7.5
safety==2.3.5

# Type stubs
types-redis==4.6.0
types-pyyaml==6.0.12.12
pandas-stubs==2.0.3.230814

# Data download
gdown==4.7.1